Improve build pipeline
Previously, we were relying on a single script that invoked TypeScript a
bunch of times to do the build.

This strategy had a few problems: it made the builds pretty slow (and
therefore made the TDD cycle slow), it also couldn't do things that we
wanted, like running tests in the browser.

This commit migrates the build process to Ember CLI, with a number of
benefits:

- We can take advantage of Ember CLI's incremental build optimizations
  to get much faster builds.
- We can split the development and production builds very easily,
  shrinking the time we need for each environment.
- The improved build times allows us to take advantage of `ember test
  -s`, which runs tests every time TypeScript files change.
- `ember test -s` also uses Testem, which allowed us to easily add
  browser-based testing (using Headless Chrome) on top of Node.

Note that changing to Ember CLI is an internal build detail, all of the
existing `npm run` commands and VSCode integration continue to work.

We also improved the VSCode integration:

- Shift-Command-B (run a build) only runs the type checker, which is
  much faster than running an entire build a few times.
- We added a "TDD" task, which opens a terminal with testem running that
  will update as files are saved.